+++ app/src/main/java/woowacourse/omok/domain/model/position/Position.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.domain.model.position
+
+data class Position(val column: Column, val row: Row) {
+    constructor(column: Int, inColumnRange: Boolean, row: Int, inRowRange: Boolean) : this(
+        Column.from(column, inColumnRange),
+        Row.from(row, inRowRange),
+    )
+}
