+++ app/src/main/java/woowacourse/omok/domain/model/position/Row.kt
@@ -1,0 +1,14 @@
+package woowacourse.omok.domain.model.position
+
+@JvmInline
+value class Row private constructor(val value: Int) {
+    companion object {
+        fun from(
+            value: Int,
+            isBoardRange: Boolean,
+        ): Row {
+            require(isBoardRange) { "잘못된 위치입니다." }
+            return Row(value)
+        }
+    }
+}
