+++ app/src/main/java/woowacourse/omok/domain/model/position/Stone.kt
@@ -1,0 +1,13 @@
+package woowacourse.omok.domain.model.position
+
+import woowacourse.omok.domain.model.stone.StoneType
+
+data class Stone(
+    val position: Position,
+    val stoneType: StoneType,
+) {
+    constructor(column: Int, inColumnRange: Boolean, row: Int, inRowRange: Boolean, stoneType: StoneType) : this(
+        Position(column, inColumnRange, row, inRowRange),
+        stoneType,
+    )
+}
