+++ app/src/main/java/woowacourse/omok/domain/model/stone/StoneType.kt
@@ -1,0 +1,13 @@
+package woowacourse.omok.domain.model.stone
+
+enum class StoneType {
+    BLACK,
+    WHITE,
+    ;
+
+    fun reverse() =
+        when (this) {
+            BLACK -> WHITE
+            WHITE -> BLACK
+        }
+}
