+++ app/src/test/java/woowacourse/omok/domain/model/BoardTest.kt
@@ -1,0 +1,26 @@
+package omok.domain.model
+
+import omok.positionOneAndOne
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import woowacourse.omok.domain.model.Board
+import woowacourse.omok.domain.model.position.Stone
+import woowacourse.omok.domain.model.stone.StoneType
+
+class BoardTest {
+    private lateinit var board: Board
+
+    @BeforeEach
+    fun setUp() {
+        board = Board()
+    }
+
+    @Test
+    fun `바둑돌을 둔다`() {
+        val position = positionOneAndOne
+        val stone = Stone(position, StoneType.BLACK)
+        board = board.addedBoard(stone)
+        assertThat(board.stones.hasStone(Stone(position, StoneType.BLACK))).isTrue()
+    }
+}
