+++ app/src/test/java/woowacourse/omok/study/MapTest.kt
