--- src/main/kotlin/omok/OmokApplication.kt
@@ -1,8 +1,0 @@
-package omok
-
-import omok.controller.OmokController
-
-fun main() {
-    val omokController = OmokController()
-    omokController.run()
-}
