--- src/main/kotlin/omok/domain/model/position/Column.kt
@@ -1,14 +1,0 @@
-package omok.domain.model.position
-
-@JvmInline
-value class Column private constructor(val value: Int) {
-    companion object {
-        fun from(
-            value: Int,
-            isBoardRange: Boolean,
-        ): Column {
-            require(isBoardRange) { "잘못된 위치입니다." }
-            return Column(value)
-        }
-    }
-}
