--- src/main/kotlin/omok/domain/model/state/Finish.kt
@@ -1,18 +1,0 @@
-package omok.domain.model.state
-
-import omok.domain.model.position.Position
-import omok.domain.model.stone.StoneType
-import omok.domain.model.stone.Stones
-
-class Finish(
-    override val stones: Stones,
-    override val stoneType: StoneType,
-) : OmokState {
-    override fun placeStone(
-        position: Position,
-        onPlaceMessage: (String) -> Unit,
-    ): OmokState {
-        onPlaceMessage("끝난 상태에서 돌을 둘 수 없습니다.")
-        return this
-    }
-}
