--- src/main/kotlin/omok/domain/model/state/OmokState.kt
@@ -1,15 +1,0 @@
-package omok.domain.model.state
-
-import omok.domain.model.position.Position
-import omok.domain.model.stone.StoneType
-import omok.domain.model.stone.Stones
-
-interface OmokState {
-    val stoneType: StoneType
-    val stones: Stones
-
-    fun placeStone(
-        position: Position,
-        onPlaceMessage: (String) -> Unit,
-    ): OmokState
-}
