--- src/main/kotlin/omok/domain/model/stone/StoneType.kt
@@ -1,13 +1,0 @@
-package omok.domain.model.stone
-
-enum class StoneType {
-    BLACK,
-    WHITE,
-    ;
-
-    fun reverse() =
-        when (this) {
-            BLACK -> WHITE
-            WHITE -> BLACK
-        }
-}
